Add files via upload
The HTML structure contains a section with a bg class that will be used as the background image container, and a div with a loading-text class that will display the loading %.

For styling: imported Ubuntu font. Set up the layout for centering the content. Defined the .bg class to style the background image and its blur effect.

JS: Initialized the load variable to 0 and started a setInterval that calls the blurring function every 30 ms. The blurring function updates the inner text of the loading text to show the current percentage.